# api/openapi.yaml
openapi: 3.1.0
info:
  title: Raincheck API
  version: 0.1.0
  description: All calendar values use UTC. Raw participant edit tokens occur only in successful join responses.
paths:
  /api/groups:
    post:
      summary: Create a UTC group without a poll
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateGroupRequest' } } } }
      responses: { '201': { $ref: '#/components/responses/CreateGroup' }, '400': { $ref: '#/components/responses/InvalidRequest' } }
  /api/groups/{inviteCode}:
    parameters: [{ $ref: '#/components/parameters/InviteCode' }]
    get:
      summary: Public group context
      responses: { '200': { $ref: '#/components/responses/PublicGroup' }, '404': { $ref: '#/components/responses/Error' } }
  /api/groups/{inviteCode}/participants:
    parameters: [{ $ref: '#/components/parameters/InviteCode' }]
    post:
      summary: Join a group
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ParticipantInput' } } } }
      responses: { '201': { $ref: '#/components/responses/Join' }, '400': { $ref: '#/components/responses/InvalidRequest' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /api/groups/{inviteCode}/participants/me:
    parameters: [{ $ref: '#/components/parameters/InviteCode' }, { $ref: '#/components/parameters/ParticipantToken' }]
    patch:
      summary: Rename the authenticated participant
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ParticipantInput' } } } }
      responses: { '200': { $ref: '#/components/responses/ParticipantResponse' }, '400': { $ref: '#/components/responses/InvalidRequest' }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /api/groups/{inviteCode}/workspace:
    parameters: [{ $ref: '#/components/parameters/InviteCode' }, { $ref: '#/components/parameters/OptionalParticipantToken' }]
    get:
      summary: Group workspace; missing, invalid, or cross-group optional token returns me null
      responses: { '200': { $ref: '#/components/responses/Workspace' }, '404': { $ref: '#/components/responses/Error' } }
  /api/groups/{inviteCode}/polls:
    parameters: [{ $ref: '#/components/parameters/InviteCode' }]
    get:
      summary: List polls newest first
      responses: { '200': { $ref: '#/components/responses/Polls' }, '404': { $ref: '#/components/responses/Error' } }
    post:
      summary: Create a poll when the group has no open poll
      parameters: [{ $ref: '#/components/parameters/ParticipantToken' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PollInput' } } } }
      responses: { '201': { $ref: '#/components/responses/PollResponse' }, '400': { $ref: '#/components/responses/InvalidRequest' }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' }, '422': { $ref: '#/components/responses/InvalidSchedule' } }
  /api/groups/{inviteCode}/polls/{pollId}/close:
    parameters: [{ $ref: '#/components/parameters/InviteCode' }, { $ref: '#/components/parameters/PollId' }, { $ref: '#/components/parameters/ParticipantToken' }]
    post:
      summary: Permanently close a poll and keep it in group history
      responses: { '200': { $ref: '#/components/responses/PollResponse' }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /api/groups/{inviteCode}/polls/{pollId}:
    parameters: [{ $ref: '#/components/parameters/InviteCode' }, { $ref: '#/components/parameters/PollId' }]
    get:
      summary: Poll configuration and participant response states
      responses: { '200': { $ref: '#/components/responses/PollDetail' }, '404': { $ref: '#/components/responses/Error' } }
...
components:
  parameters:
    InviteCode: { name: inviteCode, in: path, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_-]{32,64}$' } }
    PollId: { name: pollId, in: path, required: true, schema: { type: string, format: uuid } }
    ParticipantToken: { name: X-Participant-Token, in: header, required: true, schema: { type: string, minLength: 32, maxLength: 128 } }
    OptionalParticipantToken: { name: X-Participant-Token, in: header, required: false, schema: { type: string, maxLength: 128 } }
  responses:
    Error: { description: Stable API error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    InvalidRequest: { description: INVALID_REQUEST, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    InvalidSchedule: { description: INVALID_SCHEDULE, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    CreateGroup: { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/CreateGroupResponse' } } } }
    Join: { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/JoinResponse' } } } }
    PublicGroup: { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PublicGroupResponse' } } } }
    Workspace: { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Workspace' } } } }
    Polls: { description: OK, content: { application/json: { schema: { type: object, required: [polls], properties: { polls: { type: array, items: { $ref: '#/components/schemas/Poll' } } } } } } }
    PollResponse: { description: OK, content: { application/json: { schema: { type: object, required: [poll], properties: { poll: { $ref: '#/components/schemas/Poll' } } } } } }
    PollDetail: { description: OK, content: { application/json: { schema: { type: object, required: [poll, participantStates], properties: { poll: { $ref: '#/components/schemas/Poll' }, participantStates: { type: array } } } } } }
    ParticipantResponse: { description: OK, content: { application/json: { schema: { type: object, required: [participant], properties: { participant: { $ref: '#/components/schemas/Participant' } } } } } }
...
  schemas:
    Error: { type: object, required: [error], properties: { error: { type: object, required: [code, message, requestId], properties: { code: { enum: [INVALID_REQUEST, UNAUTHORIZED, GROUP_NOT_FOUND, POLL_NOT_FOUND, RESPONSE_NOT_FOUND, PARTICIPANT_NAME_TAKEN, RESPONSE_ALREADY_EXISTS, POLL_STATE_CONFLICT, INVALID_SCHEDULE, INTERNAL_ERROR] }, message: { type: string }, requestId: { type: string } } } } }
    Group: { type: object, required: [id, name, inviteCode, timezone], properties: { id: { type: string, format: uuid }, name: { type: string }, inviteCode: { type: string }, timezone: { const: UTC } } }
    Participant: { type: object, required: [id, displayName, avatarColor], properties: { id: { type: string, format: uuid }, displayName: { type: string }, avatarColor: { type: string, enum: [green, blue, purple, rose, yellow, gray] } } }
    ParticipantInput: { type: object, required: [displayName, avatarColor], properties: { displayName: { type: string, minLength: 1, maxLength: 80 }, avatarColor: { type: string, enum: [green, blue, purple, rose, yellow, gray] } } }
    PollInput: { type: object, required: [startsOn, endsOn, dayStart, dayEnd, slotMinutes, meetingDurationMinutes], properties: { title: { type: [string, 'null'] }, startsOn: { type: string, format: date }, endsOn: { type: string, format: date }, dayStart: { type: string }, dayEnd: { type: string }, slotMinutes: { enum: [30, 60] }, meetingDurationMinutes: { type: integer, minimum: 30, maximum: 240 } } }
    Poll: { allOf: [{ $ref: '#/components/schemas/PollInput' }, { type: object, required: [id, sequenceNo, status, basedOnPollId, createdAt, closedAt], properties: { id: { type: string, format: uuid }, sequenceNo: { type: integer }, status: { enum: [OPEN, CLOSED] }, basedOnPollId: { type: [string, 'null'], format: uuid }, createdAt: { type: string, format: date-time }, closedAt: { type: [string, 'null'], format: date-time } } }] }
...
    CreateGroupRequest: { type: object, additionalProperties: false, required: [name], properties: { name: { type: string, minLength: 1, maxLength: 120 } } }
    CreateGroupResponse: { type: object, required: [group, currentPoll], properties: { group: { $ref: '#/components/schemas/Group' }, currentPoll: { type: 'null' } } }
    JoinResponse: { type: object, required: [participant, participantEditToken], properties: { participant: { $ref: '#/components/schemas/Participant' }, participantEditToken: { type: string, writeOnly: true } } }
    PublicGroupDto: { type: object, additionalProperties: false, required: [name, timezone], properties: { name: { type: string }, timezone: { const: UTC } } }
    PublicCurrentPollDto: { type: object, additionalProperties: false, required: [id, sequenceNo, startsOn, endsOn], properties: { id: { type: string, format: uuid }, sequenceNo: { type: integer }, startsOn: { type: string, format: date }, endsOn: { type: string, format: date } } }
    PublicGroupResponse: { type: object, additionalProperties: false, required: [group, currentPoll], properties: { group: { $ref: '#/components/schemas/PublicGroupDto' }, currentPoll: { anyOf: [{ $ref: '#/components/schemas/PublicCurrentPollDto' }, { type: 'null' }] } } }
    WorkspaceParticipant: { type: object, additionalProperties: false, required: [id, displayName, avatarColor, currentPollState], properties: { id: { type: string, format: uuid }, displayName: { type: string }, avatarColor: { type: string, enum: [green, blue, purple, rose, yellow, gray] }, currentPollState: { enum: [NONE, DRAFT, CONFIRMED] } } }
    Workspace: { type: object, required: [group, me, participants, polls, currentPoll], properties: { group: { $ref: '#/components/schemas/Group' }, me: { anyOf: [{ $ref: '#/components/schemas/Participant' }, { type: 'null' }] }, participants: { type: array, items: { $ref: '#/components/schemas/WorkspaceParticipant' } }, polls: { type: array, items: { $ref: '#/components/schemas/Poll' } }, currentPoll: { anyOf: [{ $ref: '#/components/schemas/Poll' }, { type: 'null' }] } } }
...